Refactor code to delay activation in BO001 and increase timeout to 1000ms
Update the code in BO001 to delay the activation by increasing the timeout from 500ms to 1000ms. This change ensures that the activation occurs after the specified delay.

Update VARIATION value in BO003 and BO004

In BO003 and BO004, update the VARIATION value from "1" to "2" in the shared.js file. This change reflects the updated variation for the Moreconversion Bloxboom client.

Add observeDOM function in BO004

In BO004, add the observeDOM function to the experiment.js file. This function observes changes in the DOM and triggers the init function when

Add badge for V1 & V2 in BO004

In BO004, add a badge component and insert it before the largestBundleButton if it doesn't already exist. This change adds a badge for V1 and V2 variations in the Moreconversion Bloxboom client.